Textures get import straight into Blender
I want to go ahead and write support for normals. cause right now, the
meshes are inside out. Textures do not yet show transparency in the
viewport, but I will try to change that.

I will also be including an option for importing alpha for rendering or
for game engine (you have to use different settings for each).

Blender also has a hard time with dispplaying the textures the right way
because tecnically, the meshes are in the same space. Ill figure out a
work around later.